requirements.txt
websockets>=11.0
psutil>=5.9
PyJWT>=2.8.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
